support for bincode
