+++ module-system/sov-state/Cargo.toml
@@ -14,6 +14,7 @@ resolver = "2"
 [dependencies]
 anyhow = { workspace = true }
 borsh = { workspace = true }
+bincode = { workspace = true }
 serde = { workspace = true }
 thiserror = { workspace = true }
 sov-rollup-interface = { path = "../../rollup-interface", version = "0.1" }
